Move TrajectoryPostProcessor.hpp into root planner directory and Smooth/Retime PostProcessor classes into existing planner/parabolic files. Remove planner/postprocessor.
