+++ include/aikido/planner/parabolic/ParabolicSmoother.hpp
@@ -3,6 +3,7 @@
 
 #include <Eigen/Dense>
 #include "aikido/constraint/Testable.hpp"
+#include "aikido/planner/TrajectoryPostProcessor.hpp"
 #include "aikido/trajectory/Interpolated.hpp"
 #include "aikido/trajectory/Spline.hpp"
 
@@ -128,6 +129,63 @@ std::unique_ptr<trajectory::Spline> doShortcutAndBlend(
     int _blendIterations = DEFAULT_BLEND_ITERATIONS,
     double _checkResolution = DEFAULT_CHECK_RESOLUTION,
     double _tolerance = DEFAULT_TOLERANCE);
+
+/// Class for performing parabolic smoothing on trajectories
+class SmoothTrajectoryPostProcessor
+    : public aikido::planner::TrajectoryPostProcessor
+{
+public:
+  /// \param _space pointer to statespace trajectories correspond to.
+  /// \param _smootherFeasibilityCheckResolution the resolution in discretizing
+  /// a segment in checking the feasibility of the segment.
+  /// \param _smootherFeasibilityApproxTolerance this tolerance is used in a
+  /// piecewise linear discretization that deviates no more than
+  /// \c _smootherFeasibilityApproxTolerance from the parabolic ramp along any
+  /// axis, and then checks for configuration and segment feasibility along that
+  /// piecewise linear path.
+  /// \param _velocityLimits maximum velocity for each dimension.
+  /// \param _accelerationLimits maximum acceleration for each dimension.
+  /// \param _collisionTestable Check whether a position is feasible.
+  /// \param _enableShortcut Whether shortcutting is used in smoothing.
+  /// \param _enableBlend Whether blending is used in smoothing.
+  /// \param _shortcutTimelimit Timelimit for shortcutting. It is ineffective
+  /// when _enableShortcut is false.
+  /// \param _blendRadius Blend radius for blending. It is ineffective
+  /// when _enableBlend is false.
+  /// \param _blendIterations Blend iterations for blending. It is
+  /// ineffective when _enableBlend is false.
+  SmoothTrajectoryPostProcessor(
+      aikido::statespace::StateSpacePtr _space,
+      const Eigen::VectorXd& _velocityLimits,
+      const Eigen::VectorXd& _accelerationLimits,
+      const aikido::constraint::TestablePtr& _collisionTestable,
+      bool _enableShortcut = true,
+      bool _enableBlend = true,
+      double _shortcutTimelimit = DEFAULT_TIMELIMT,
+      double _blendRadius = DEFAULT_BLEND_RADIUS,
+      int _blendIterations = DEFAULT_BLEND_ITERATIONS,
+      double _smootherFeasibilityCheckResolution = DEFAULT_CHECK_RESOLUTION,
+      double _smootherFeasibilityApproxTolerance = DEFAULT_TOLERANCE);
+
+  // Documentation inherited.
+  std::unique_ptr<aikido::trajectory::Spline> postprocess(
+      const aikido::trajectory::InterpolatedPtr& _inputTraj,
+      const aikido::common::RNG* _rng) override;
+
+private:
+  aikido::statespace::StateSpacePtr mSpace;
+  double mSmootherFeasibilityCheckResolution;
+  double mSmootherFeasibilityApproxTolerance;
+  const Eigen::VectorXd mVelocityLimits;
+  const Eigen::VectorXd mAccelerationLimits;
+  aikido::constraint::TestablePtr mCollisionTestable;
+
+  bool mEnableShortcut;
+  bool mEnableBlend;
+  double mShortcutTimelimit;
+  double mBlendRadius;
+  int mBlendIterations;
+};
 
 } // namespace parabolic
 } // namespace planner
