+++ src/planner/CMakeLists.txt
@@ -37,5 +37,4 @@ format_add_sources(${sources})
 
 add_subdirectory("ompl")      # [constraint], [distance], [statespace], [trajectory], dart, ompl
 add_subdirectory("parabolic") # [external], [common], [trajectory], [statespace], dart
-add_subdirectory("postprocessor") # [external], [common], [trajectory], [statespace], dart
 add_subdirectory("vectorfield") # [common], [trajectory], [statespace], dart
