+++ src/planner/parabolic/ParabolicSmoother.cpp
@@ -3,6 +3,7 @@
 #include <dart/common/StlHelpers.hpp>
 #include <aikido/common/Spline.hpp>
 #include <aikido/planner/parabolic/ParabolicSmoother.hpp>
+#include <aikido/planner/parabolic/ParabolicTimer.hpp>
 #include "DynamicPath.h"
 #include "HauserParabolicSmootherHelpers.hpp"
 #include "ParabolicUtil.hpp"
@@ -109,6 +110,92 @@ std::unique_ptr<trajectory::Spline> doShortcutAndBlend(
       = detail::convertToSpline(*dynamicPath, startTime, stateSpace);
 
   return outputTrajectory;
+}
+
+SmoothTrajectoryPostProcessor::SmoothTrajectoryPostProcessor(
+    aikido::statespace::StateSpacePtr _space,
+    const Eigen::VectorXd& _velocityLimits,
+    const Eigen::VectorXd& _accelerationLimits,
+    const aikido::constraint::TestablePtr& _collisionTestable,
+    bool _enableShortcut,
+    bool _enableBlend,
+    double _shortcutTimelimit,
+    double _blendRadius,
+    int _blendIterations,
+    double _smootherFeasibilityCheckResolution,
+    double _smootherFeasibilityApproxTolerance)
+  : mSpace{std::move(_space)}
+  , mSmootherFeasibilityCheckResolution{_smootherFeasibilityCheckResolution}
+  , mSmootherFeasibilityApproxTolerance{_smootherFeasibilityApproxTolerance}
+  , mVelocityLimits{_velocityLimits}
+  , mAccelerationLimits{_accelerationLimits}
+  , mCollisionTestable{_collisionTestable}
+  , mEnableShortcut{_enableShortcut}
+  , mEnableBlend{_enableBlend}
+  , mShortcutTimelimit{_shortcutTimelimit}
+  , mBlendRadius{_blendRadius}
+  , mBlendIterations{_blendIterations}
+{
+  // Do nothing
+}
+
+std::unique_ptr<aikido::trajectory::Spline>
+SmoothTrajectoryPostProcessor::postprocess(
+    const aikido::trajectory::InterpolatedPtr& _inputTraj,
+    const aikido::common::RNG* _rng)
+{
+  if (!_rng)
+    throw std::invalid_argument(
+        "Passed nullptr _rng to SmoothTrajectoryPostProcessor::postprocess");
+  if (!_inputTraj)
+    throw std::invalid_argument(
+        "Passed nullptr _inputTraj to "
+        "SmoothTrajectoryPostProcessor::postprocess");
+
+  // Get timed trajectory for arm
+  auto timedTrajectory = computeParabolicTiming(
+      *_inputTraj, mVelocityLimits, mAccelerationLimits);
+
+  if (mEnableShortcut && mEnableBlend)
+  {
+    return doShortcutAndBlend(
+        *timedTrajectory,
+        mCollisionTestable,
+        mVelocityLimits,
+        mAccelerationLimits,
+        *_rng->clone(),
+        mShortcutTimelimit,
+        mBlendRadius,
+        mBlendIterations,
+        mSmootherFeasibilityCheckResolution,
+        mSmootherFeasibilityApproxTolerance);
+  }
+  else if (mEnableShortcut)
+  {
+    return doShortcut(
+        *timedTrajectory,
+        mCollisionTestable,
+        mVelocityLimits,
+        mAccelerationLimits,
+        *_rng->clone(),
+        mShortcutTimelimit,
+        mSmootherFeasibilityCheckResolution,
+        mSmootherFeasibilityApproxTolerance);
+  }
+  else if (mEnableBlend)
+  {
+    return doBlend(
+        *timedTrajectory,
+        mCollisionTestable,
+        mVelocityLimits,
+        mAccelerationLimits,
+        mBlendRadius,
+        mBlendIterations,
+        mSmootherFeasibilityCheckResolution,
+        mSmootherFeasibilityApproxTolerance);
+  }
+
+  return timedTrajectory;
 }
 
 } // namespace parabolic
