+++ src/planner/parabolic/ParabolicTimer.cpp
@@ -113,6 +113,31 @@ std::unique_ptr<aikido::trajectory::Spline> computeParabolicTiming(
   return outputTrajectory;
 }
 
+RetimeTrajectoryPostProcessor::RetimeTrajectoryPostProcessor(
+    aikido::statespace::StateSpacePtr _space,
+    const Eigen::VectorXd& _velocityLimits,
+    const Eigen::VectorXd& _accelerationLimits)
+  : mSpace{std::move(_space)}
+  , mVelocityLimits{_velocityLimits}
+  , mAccelerationLimits{_accelerationLimits}
+{
+  // Do nothing
+}
+
+std::unique_ptr<aikido::trajectory::Spline>
+RetimeTrajectoryPostProcessor::postprocess(
+    const aikido::trajectory::InterpolatedPtr& _inputTraj,
+    const aikido::common::RNG* /*_rng*/)
+{
+  if (!_inputTraj)
+    throw std::invalid_argument(
+        "Passed nullptr _inputTraj to "
+        "RetimeTrajectoryPostProcessor::postprocess");
+
+  return computeParabolicTiming(
+      *_inputTraj, mVelocityLimits, mAccelerationLimits);
+}
+
 } // namespace parabolic
 } // namespace planner
 } // namespace aikido
